+++ src/components/Sparkline/stories/Sparkline.stories.tsx
@@ -50,7 +50,11 @@ export default {
 } as Meta;
 
 const Template: Story<SparklineProps> = (args: SparklineProps) => {
-  return <Sparkline {...args} />;
+  return (
+    <div style={{width: '200px', height: '100px'}}>
+      <Sparkline {...args} />
+    </div>
+  );
 };
 
 export const Default = Template.bind({});
@@ -63,4 +67,49 @@ hasSpline.args = {
   series,
   isAnimated: true,
   hasSpline: true,
+};
+
+export const OffsetAndNulls = Template.bind({});
+OffsetAndNulls.args = {
+  series: [
+    {
+      color: 'darkModePositive',
+      areaStyle: 'gradient',
+      hasPoint: true,
+      offsetRight: 12,
+      offsetLeft: 50,
+      data: [
+        {x: 0, y: 100},
+        {x: 1, y: 50},
+        {x: 2, y: null},
+        {x: 3, y: 200},
+        {x: 4, y: 400},
+        {x: 5, y: 1000},
+        {x: 6, y: 200},
+        {x: 7, y: 800},
+        {x: 8, y: 900},
+        {x: 9, y: 200},
+        {x: 10, y: 100},
+      ],
+    },
+    {
+      color: 'darkModePositive',
+      areaStyle: 'none',
+      lineStyle: 'dashed',
+      data: [
+        {x: 0, y: 20},
+        {x: 1, y: 20},
+        {x: 2, y: 20},
+        {x: 3, y: 20},
+        {x: 4, y: 20},
+        {x: 5, y: 20},
+        {x: 6, y: 20},
+        {x: 7, y: 20},
+        {x: 8, y: 20},
+        {x: 9, y: 20},
+        {x: 10, y: 20},
+      ],
+    },
+  ],
+  isAnimated: true,
 };